requirements.txt
aiogram==3.27.0
beautifulsoup4==4.14.3
fanficfare==4.56.0
lxml==6.1.0
python-dotenv==1.2.2
